// src/app/models/flex-workflows-model.ts
export class flexworkflowmodel {
  Id: number;
  Name: string;
  UILabel: string;
  ScopeEntityType: string;
  ScopeEntityId: number;
  EntityType: string;
  EntityId: number;
  Description: string;
  CreatedById: number;
  UpdatedById: number;
  CreatedBy: string;
  UpdatedBy: string;
  CreatedDateTime: Date;
  UpdatedDdateTime: Date;
  IsPortfolioEnabled: boolean;
  IsManualBranchSelection: boolean;
  ParameterCarrier: string;
  ProjectName: string;
  Status: string;
  DiagramJson: string;
  LoginUserRoleId: number;
  HelpFilePath: string;
  SelectedTagIdList: string;
}